create domains, routes and navigation
