+++ src/pages/About.jsx
@@ -1,0 +1,13 @@
+
+
+export default function About (){
+
+
+    return (
+        <>
+        This is about
+        </>
+
+
+    );
+}
