+++ src/pages/Error.jsx
@@ -1,0 +1,13 @@
+
+
+export default function Error (){
+
+
+    return (
+        <>
+        This is error
+        </>
+
+
+    );
+}
