+++ src/pages/FakeData/FakeData.jsx
@@ -1,0 +1,11 @@
+import CreateFakeOrders from "./CreateFakeOrders.jsx";
+
+export default function FakeData() {
+  return (
+    <>
+      <div class="w3-container">
+        <CreateFakeOrders />
+      </div>
+    </>
+  );
+}
