+++ src/pages/Home.jsx
@@ -1,0 +1,13 @@
+
+
+export default function Home (){
+
+
+    return (
+        <>
+        This is home
+        </>
+
+
+    );
+}
