+++ src/pages/Orders/Orders.jsx
@@ -1,0 +1,11 @@
+import OrdersList from "./OrdersList.jsx";
+
+export default function Orders() {
+  return (
+    <>
+      <div class="w3-container w3-light-grey">
+        <OrdersList />
+      </div>
+    </>
+  );
+}
