+++ src/pages/Tables/Tables.jsx
@@ -1,0 +1,11 @@
+import OrdersTable from "./OrdersTable.jsx";
+
+export default function Tables() {
+  return (
+    <>
+      <div class="w3-container">
+        <OrdersTable />
+      </div>
+    </>
+  );
+}
